fix: score range and precision

+++ src/main/java/com/project/api/services/ScoreService.java
@@ -12,7 +12,6 @@ import jakarta.transaction.Transactional;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
-
 @Service
 public class ScoreService {
 
@@ -27,6 +26,7 @@ public class ScoreService {
 
     @Transactional
     public MovieDTO saveScore(ScoreDTO dto) {
+
         User user = userRepository.findByEmail(dto.getEmail());
         if(user == null) {
             user = new User();
@@ -44,12 +44,12 @@ public class ScoreService {
 
         scoreRepository.saveAndFlush(score);
 
-        double sum = 0.0;
+        float sum = 0;
         for (Score s: movie.getScores() ) {
             sum = sum + s.getValue();
         }
 
-        double avg = sum / movie.getScores().size();
+        float avg = sum / movie.getScores().size();
 
         movie.setScore(avg);
         movie.setCount(movie.getScores().size());
